docs: LSDV-5343: Docs for visibleWhen; remove FF_DEV_1372 (#1483)
Document `visibleWhen` and connected `when*` properties to highlight their depencies on each other. Even if some nuances will be improved and validated in the future, these nuances right now will help users a lot to understand problems with configs.

Basically, `visibleWhen` is required parameter, can refer to choices or regions, both selected or unselected:
- region-selected, can specify tag and label value, both are optional
- choice-selected, can specify tag, and with tag specified can specify choice value
- no-region-selected, can't specify anything else
- choice-unselected, can specify tag, and with tag specified can specify choice value
Can be applied to `View` and `Choices` tags.

Also small FF_DEV_1372 was removed because it's enabled everywhere. And documentation for `choice-unselected` option by this FF was added to docs.

* Remove "choice-unselected" FF (FF_DEV_1372); fix docs

* Remove wrongly used whenChoiceValue

* Improve docs to highlight params dependency

* Only docs in this PR; extract fix

Extract fix to different PR and wrap it in FF just in case

* Remove FF_DEV_1372 from import

* Remove FF_DEV_1372 from the list

* Remove flag from tests as well

---------

Co-authored-by: hlomzik <hlomzik@users.noreply.github.com>